Feat: 컴포넌트 이름 수정
